Refactoring: Move MapLegend into Map-Scope
This whole folder was only used by the Map-"Product" (HBI/Planungen) (by now), so we should move it in the component folder over there.

Also import styles directly from the styles file and rename that so it's easier to understand the imports.
