+++ src/apps/Map/components/MapLegend/CadastreLegendGrid.tsx
@@ -13,7 +13,7 @@ import {
   LegendCol,
   LegendItem,
   LegendWrapper,
-} from './index';
+} from './Styles';
 
 export const CadastreLegendGrid = () => (
   <LegendWrapper>
