+++ src/apps/Map/components/MapLegend/Styles.tsx
