+++ src/apps/Map/components/MapLegend/assets/construction.png
