+++ src/apps/Map/components/MapLegend/assets/dividers.png
