+++ src/apps/Map/components/MapLegend/assets/footways.png
