+++ src/apps/Map/components/MapLegend/assets/guardrail-legend.png
