+++ src/apps/Map/components/MapLegend/index.ts
@@ -1,4 +1,3 @@
 export * from './Legend';
 export * from './IncidentLegendGrid';
 export * from './CadastreLegendGrid';
-export * from './LegendStyles';
